web_app/requirements.txt
-i https://pypi.org/simple

email-validator==2.2.0; python_version >= '3.8'
flask==3.1.0; python_version >= '3.9'
flask-bcrypt==1.0.1
flask-login==0.6.3; python_version >= '3.7'
flask-pymongo==3.0.1; python_version >= '3.9'
flask-wtf==1.2.2; python_version >= '3.9'
pandas==2.2.3; python_version >= '3.9'
pillow==11.2.1; python_version >= '3.9'
pillow-heif==0.22.0; python_version >= '3.9'
boto3==1.34.43; python_version >= '3.9'
python-dotenv==1.1.0; python_version >= '3.9'
speciesnet==4.0.3; python_version < '3.12' and python_version >= '3.9'
gunicorn
